Update modifier keys to be "Cmd+Ctrl"
Using "Alt" has the nasty side-effect that if you move or resize
a window that is not in focus, the window you are switching away
from gets hidden.

This modifier combo was switched to "Alt" before release to
guarantee that this could be run with Afloat.  Reverting to
the original combo means that we'll have to settle for
being "pretty sure" that we're compatible (as long
as mac-move-resize loads after Afloat, we're
fine.  This always *seems* to be the case, but there
doesn't seem to be any code in SIMBL which guarantees the
order)